requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.3.0
googlemaps==4.9.0
pandas==2.0.3
python-docx==0.8.11
pdfplumber==0.9.0
PyMuPDF==1.23.7
geopy==2.3.0
requests==2.31.0
setuptools==68.2.2 